requirements.txt
cassandra-driver
ccm
